Add flag to append proxy host to forwarded headers
If there are multiple proxying steps the host name of this vhost may not
be the only forwarded host name. This flags allows the user to append
the host name onto an incomming X-Forwareded-{Server,Host} header rather
than setting it.

Ideally the HttpProxyModule would provide variables that handle this
(like `$proxy_add_x_forwarded_for`) but it doesn't.